% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicluster.R
\name{print.summary.multicluster}
\alias{print.summary.multicluster}
\title{Print summary of multicluster results}
\usage{
\method{print}{summary.multicluster}(x, ...)
}
\arguments{
\item{x}{An summary.multicluster object to be printed.}

\item{...}{Ignored.}
}
\description{
Print summary of multicluster results to the terminal.
}
\seealso{
\link{spatial.scan}, \link{summary.multicluster}
}
