% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicluster.R
\name{spatial.scan}
\alias{spatial.scan}
\title{Detect multiple spatial clusters using the flexible/circular scan statistic}
\usage{
spatial.scan(
  x,
  y,
  name,
  observed,
  expected,
  nb,
  clustersize = 15,
  maxclusters = 25,
  stattype = "ORIGINAL",
  scanmethod = "FLEXIBLE",
  ralpha = 0.2,
  simcount = 999,
  verbose = FALSE,
  cores = 2
)
}
\arguments{
\item{x}{A vector of X-coordinates.}

\item{y}{A vector of Y-coordinates.}

\item{name}{A vector of names of each area.}

\item{observed}{A vector with the observed number of disease cases.}

\item{expected}{A vector with the expected number of disease cases under the null hypothesis.
This is used on "Poisson" model.}

\item{nb}{A neighbors list or an adjacency matrix.}

\item{clustersize}{The number of maximum spatial cluster size to scan, i.e., the maximum number
of regions included in the detected cluster}

\item{maxclusters}{The maximum number of clusters to be detected.}

\item{stattype}{Statistic type to be used (case-insensitive).
\describe{
  \item{"ORIGINAL"}{the likelihood ratio statistic by Kulldorff and
  Nagarwalla (1995)}
  \item{"RESTRICTED"}{the restricted likelihood ratio statistic by Tango
  (2008), with a preset parameter \code{ralpha} for restriction}
}}

\item{scanmethod}{Scanning method to be used (case-insensitive).
\describe{
  \item{"FLEXIBLE"}{flexible scan statistic by Tango and Takahashi (2005)}
  \item{"CIRCULAR"}{circular scan statistic by Kulldorff (1997)}
}}

\item{ralpha}{Threshold parameter of the middle p-value for the restricted likelihood ratio
statistic.}

\item{simcount}{The number of Monte Carlo replications to calculate a p-value for statistical
test.}

\item{verbose}{Print progress messages.}

\item{cores}{The number of CPU cores used for Monte Carlo replications.}
}
\value{
An \code{rflexscan} object which contains analysis results and specified
parameters.
}
\description{
This function analyzes spatial count data using the flexible spatial scan
statistic developed by Tango and Takahashi (2005) or Kulldorff's circular
spatial scan statistic (1997), and detect spatial disease clusters.
}
\details{
Centroid coordinates for each region should be specified EITHER by Cartesian
coordinates using arguments \code{x} and \code{y} or by latitudes and
longitudes using arguments \code{lat} and \code{lon}.
Note that \code{lat} and \code{lon} are DEPRECATED due to accuracy issues.
This feature is implemented to maintain compatibility with FleXScan software.
We recommend to transform latitude and longitude onto the Cartesian
coordinate system beforehand (using \code{spTransform} function in sp package,
for example) and use the \code{x} and \code{y} parameters that are projected
coordinates.
}
\references{
Tango T. and Takahashi K. (2005). A flexibly shaped spatial scan
  statistic for detecting clusters, International Journal of Health
  Geographics 4:11.

  Kulldorff M. and Nagarwalla N. (1995). Spatial disease clusters:
  Detection and Inference. Statistics in Medicine 14:799-810.

  Kulldorff M. (1997). A spatial scan statistic. Communications in
  Statistics: Theory and Methods, 26:1481-1496.

  Tango T. (2008). A spatial scan statistic with a restricted
  likelihood ratio. Japanese Journal of Biometrics 29(2):75-95.
}
