% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicluster.R
\name{summary.multicluster}
\alias{summary.multicluster}
\title{Summarizing multicluster results}
\usage{
\method{summary}{multicluster}(object, ...)
}
\arguments{
\item{object}{An rflexscan object to be summarized.}

\item{...}{Ignored.}
}
\description{
Summary method for multicluster objects.
}
\seealso{
\link{spatial.scan}
}
